payment system upgradeed, order cancel system added

+++ src/Pages/Dashboard/UserRow.js
@@ -20,19 +20,23 @@ const UserRow = ({ user, refetch, index }) => {
     }
 
     const handleDelete = email => {
-        fetch(`http://localhost:5000/user/${email}`, {
-            method: 'DELETE',
-            headers: {
-                'authorization': `Bearer ${localStorage.getItem('accessToken')}`
-            },
-        })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount) {                   
-                    toast.success('Successfully added an Admin')
-                    refetch();
-                }
+        const proceed = window.confirm('Are You Sure?')
+        if (proceed){
+            fetch(`http://localhost:5000/user/${email}`, {
+                method: 'DELETE',
+                headers: {
+                    'authorization': `Bearer ${localStorage.getItem('accessToken')}`
+                },
             })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount) {
+                        toast.success('Successfully deleted an User')
+                        refetch();
+                    }
+                })
+        }
+
     }
 
     return (
@@ -41,7 +45,9 @@ const UserRow = ({ user, refetch, index }) => {
             <td>{user.email}</td>
             <td>{role !== 'admin' && <button onClick={makeAdmin} class="btn btn-success text-white hover:bg-green-600 border-0 btn-xs"> <svg xmlns="http://www.w3.org/2000/svg" class="stroke-current flex-shrink-0 h-6 w-6" fill="none" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z" /></svg> Make Admin</button>}
             </td>
-            <td><button onClick={()=> handleDelete(email)} class="btn btn-error text-white hover:bg-red-600 border-0 btn-xs"> <svg xmlns="http://www.w3.org/2000/svg" class="stroke-current flex-shrink-0 h-6 w-6" fill="none" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 14l2-2m0 0l2-2m-2 2l-2-2m2 2l2 2m7-2a9 9 0 11-18 0 9 9 0 0118 0z" /></svg> Remove User</button></td>
+            <td><button onClick={() => handleDelete(email)} class="btn btn-error text-white hover:bg-red-600 border-0 btn-xs"> <svg xmlns="http://www.w3.org/2000/svg" class="stroke-current flex-shrink-0 h-6 w-6" fill="none" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 14l2-2m0 0l2-2m-2 2l-2-2m2 2l2 2m7-2a9 9 0 11-18 0 9 9 0 0118 0z" /></svg> Remove User</button>
+
+            </td>
         </tr>
     );
 };
